Moved GeoNodePublishHandler: handlers.__init__ -> handlers.import_handler.  Added handle() test.
